COMMERCE-9437 Rename
